# App.API/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"Company": {
				"properties": {
					"id": {
						"type": "string",
						"example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
					},
					"version": {
						"type": "number",
						"format": "double",
						"example": 1
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T10:00:00Z"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T11:30:00Z"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T12:00:00Z"
					},
					"createdByUserId": {
						"type": "string",
						"example": "u1s2e3r4-i5d6-7890-1234-567890abcdef"
					},
					"updatedByUserId": {
						"type": "string",
						"example": "u1s2e3r4-i5d6-7890-1234-567890abcdef"
					},
					"name": {
						"type": "string",
						"example": "Acme Corp"
					},
					"timezone": {
						"type": "string",
						"example": "America/New_York"
					},
					"users": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					},
					"teams": {
						"items": {
							"$ref": "#/components/schemas/Team"
						},
						"type": "array"
					},
					"actionCodes": {
						"items": {
							"$ref": "#/components/schemas/ActionCode"
						},
						"type": "array"
					},
					"timesheetEntries": {
						"items": {
							"$ref": "#/components/schemas/TimesheetEntry"
						},
						"type": "array"
					},
					"teamMembers": {
						"items": {
							"$ref": "#/components/schemas/TeamMember"
						},
						"type": "array"
					},
					"timesheetHistory": {
						"items": {
							"$ref": "#/components/schemas/TimesheetHistory"
						},
						"type": "array"
					},
					"companySettings": {
						"$ref": "#/components/schemas/CompanySettings"
					},
					"menuCategories": {
						"items": {
							"$ref": "#/components/schemas/MenuCategory"
						},
						"type": "array"
					},
					"menuCards": {
						"items": {
							"$ref": "#/components/schemas/MenuCard"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"version",
					"createdAt",
					"updatedAt",
					"name",
					"users",
					"teams",
					"actionCodes",
					"timesheetEntries",
					"teamMembers",
					"timesheetHistory",
					"companySettings",
					"menuCategories",
					"menuCards"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Role": {
				"properties": {
					"id": {
						"type": "string",
						"example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
					},
					"version": {
						"type": "number",
						"format": "double",
						"example": 1
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T10:00:00Z"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T11:30:00Z"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T12:00:00Z"
					},
					"createdByUserId": {
						"type": "string",
						"example": "u1s2e3r4-i5d6-7890-1234-567890abcdef"
					},
					"updatedByUserId": {
						"type": "string",
						"example": "u1s2e3r4-i5d6-7890-1234-567890abcdef"
					},
					"companyId": {
						"type": "string",
						"example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
					},
					"company": {
						"$ref": "#/components/schemas/Company"
					},
					"name": {
						"type": "string",
						"example": "Admin"
					},
					"description": {
						"type": "string",
						"example": "Administrator with full access to all company resources."
					},
					"rolePermissions": {
						"items": {
							"$ref": "#/components/schemas/RolePermission"
						},
						"type": "array"
					},
					"users": {
						"items": {
							"$ref": "#/components/schemas/User"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"version",
					"createdAt",
					"updatedAt",
					"companyId",
					"company",
					"name",
					"rolePermissions",
					"users"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"User": {
				"properties": {
					"id": {
						"type": "string",
						"example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
					},
					"version": {
						"type": "number",
						"format": "double",
						"example": 1
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T10:00:00Z"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T11:30:00Z"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T12:00:00Z"
					},
					"createdByUserId": {
						"type": "string",
						"example": "u1s2e3r4-i5d6-7890-1234-567890abcdef"
					},
					"updatedByUserId": {
						"type": "string",
						"example": "u1s2e3r4-i5d6-7890-1234-567890abcdef"
					},
					"companyId": {
						"type": "string",
						"example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
					},
					"company": {
						"$ref": "#/components/schemas/Company"
					},
					"email": {
						"type": "string",
						"example": "john.doe@example.com"
					},
					"firstName": {
						"type": "string",
						"example": "John"
					},
					"lastName": {
						"type": "string",
						"example": "Doe"
					},
					"passwordHash": {
						"type": "string"
					},
					"mustChangePasswordAtNextLogin": {
						"type": "boolean",
						"example": false
					},
					"roleId": {
						"type": "string",
						"example": "r1o2l3e4-i5d6-7890-1234-567890abcdef"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"phoneNumber": {
						"type": "string",
						"example": "+15551234567"
					},
					"lastLogin": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T11:30:00Z"
					},
					"isAnonymized": {
						"type": "boolean",
						"example": false
					},
					"activeSessions": {
						"items": {
							"$ref": "#/components/schemas/ActiveSession"
						},
						"type": "array"
					},
					"statusId": {
						"type": "string",
						"example": "s1t2a3t4-u5s6-7890-1234-567890abcdef"
					},
					"status": {
						"$ref": "#/components/schemas/UserStatus"
					}
				},
				"required": [
					"id",
					"version",
					"createdAt",
					"updatedAt",
					"companyId",
					"company",
					"email",
					"firstName",
					"lastName",
					"passwordHash",
					"mustChangePasswordAtNextLogin",
					"roleId",
					"role",
					"isAnonymized",
					"activeSessions",
					"statusId",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Team": {
				"properties": {
					"id": {
						"type": "string",
						"example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
					},
					"version": {
						"type": "number",
						"format": "double",
						"example": 1
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T10:00:00Z"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T11:30:00Z"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T12:00:00Z"
					},
					"createdByUserId": {
						"type": "string",
						"example": "u1s2e3r4-i5d6-7890-1234-567890abcdef"
					},
					"updatedByUserId": {
						"type": "string",
						"example": "u1s2e3r4-i5d6-7890-1234-567890abcdef"
					},
					"companyId": {
						"type": "string",
						"example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
					},
					"company": {
						"$ref": "#/components/schemas/Company"
					},
					"name": {
						"type": "string",
						"example": "Engineering"
					},
					"members": {
						"items": {
							"$ref": "#/components/schemas/TeamMember"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"version",
					"createdAt",
					"updatedAt",
					"companyId",
					"company",
					"name",
					"members"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TeamMember": {
				"properties": {
					"id": {
						"type": "string",
						"example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
					},
					"version": {
						"type": "number",
						"format": "double",
						"example": 1
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T10:00:00Z"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T11:30:00Z"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T12:00:00Z"
					},
					"createdByUserId": {
						"type": "string",
						"example": "u1s2e3r4-i5d6-7890-1234-567890abcdef"
					},
					"updatedByUserId": {
						"type": "string",
						"example": "u1s2e3r4-i5d6-7890-1234-567890abcdef"
					},
					"companyId": {
						"type": "string",
						"example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
					},
					"company": {
						"$ref": "#/components/schemas/Company"
					},
					"teamId": {
						"type": "string",
						"example": "g1h2i3j4-k5l6-7890-1234-567890abcdef"
					},
					"team": {
						"$ref": "#/components/schemas/Team"
					},
					"userId": {
						"type": "string",
						"example": "u1s2e3r4-i5d6-7890-1234-567890abcdef"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"role": {
						"type": "string",
						"example": "member"
					}
				},
				"required": [
					"id",
					"version",
					"createdAt",
					"updatedAt",
					"companyId",
					"company",
					"teamId",
					"team",
					"userId",
					"user",
					"role"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ActionCode": {
				"properties": {
					"id": {
						"type": "string",
						"example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
					},
					"version": {
						"type": "number",
						"format": "double",
						"example": 1
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T10:00:00Z"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T11:30:00Z"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T12:00:00Z"
					},
					"createdByUserId": {
						"type": "string",
						"example": "u1s2e3r4-i5d6-7890-1234-567890abcdef"
					},
					"updatedByUserId": {
						"type": "string",
						"example": "u1s2e3r4-i5d6-7890-1234-567890abcdef"
					},
					"companyId": {
						"type": "string",
						"example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
					},
					"company": {
						"$ref": "#/components/schemas/Company"
					},
					"code": {
						"type": "string",
						"example": "DEV"
					},
					"name": {
						"type": "string",
						"example": "Development"
					},
					"color": {
						"type": "string",
						"example": "#FF5733"
					},
					"type": {
						"$ref": "#/components/schemas/ActionCodeType",
						"example": "billable"
					},
					"active": {
						"type": "boolean",
						"example": true
					},
					"billableDefault": {
						"$ref": "#/components/schemas/ActionCodeBillableDefault"
					},
					"billableEditable": {
						"type": "boolean"
					},
					"locationPolicy": {
						"$ref": "#/components/schemas/ActionCodeLocationPolicy"
					}
				},
				"required": [
					"id",
					"version",
					"createdAt",
					"updatedAt",
					"companyId",
					"company",
					"code",
					"name",
					"color",
					"type",
					"active",
					"billableDefault",
					"billableEditable",
					"locationPolicy"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TimesheetEntry": {
				"properties": {
					"id": {
						"type": "string",
						"example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
					},
					"version": {
						"type": "number",
						"format": "double",
						"example": 1
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T10:00:00Z"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T11:30:00Z"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T12:00:00Z"
					},
					"createdByUserId": {
						"type": "string",
						"example": "u1s2e3r4-i5d6-7890-1234-567890abcdef"
					},
					"updatedByUserId": {
						"type": "string",
						"example": "u1s2e3r4-i5d6-7890-1234-567890abcdef"
					},
					"companyId": {
						"type": "string",
						"example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
					},
					"company": {
						"$ref": "#/components/schemas/Company"
					},
					"userId": {
						"type": "string",
						"example": "u1s2e3r4-i5d6-7890-1234-567890abcdef"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"timesheetId": {
						"type": "string",
						"example": "t1i2m3e4-s5h6e7e8-9012-3456-7890abcdef"
					},
					"timesheet": {
						"$ref": "#/components/schemas/Timesheet"
					},
					"timesheetRowId": {
						"type": "string"
					},
					"timesheetRow": {
						"$ref": "#/components/schemas/TimesheetRow"
					},
					"actionCodeId": {
						"type": "string",
						"example": "a1c2t3i4-o5n6-7890-1234-567890abcdef"
					},
					"actionCode": {
						"$ref": "#/components/schemas/ActionCode"
					},
					"workMode": {
						"$ref": "#/components/schemas/WorkMode",
						"example": "office"
					},
					"country": {
						"type": "string",
						"example": "US"
					},
					"startedAt": {
						"type": "string",
						"format": "date-time",
						"example": "2024-01-01T09:00:00Z"
					},
					"endedAt": {
						"type": "string",
						"format": "date-time",
						"example": "2024-01-01T17:00:00Z"
					},
					"durationMin": {
						"type": "number",
						"format": "double",
						"example": 480
					},
					"day": {
						"type": "string",
						"example": "2024-01-01"
					},
					"note": {
						"type": "string",
						"example": "Worked on feature X"
					},
					"status": {
						"$ref": "#/components/schemas/TimesheetEntryStatus",
						"example": "saved"
					},
					"statusUpdatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"version",
					"createdAt",
					"updatedAt",
					"companyId",
					"company",
					"userId",
					"user",
					"actionCodeId",
					"actionCode",
					"workMode",
					"country",
					"durationMin",
					"day",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"TimesheetHistory": {
				"properties": {
					"id": {
						"type": "string",
						"example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
					},
					"version": {
						"type": "number",
						"format": "double",
						"example": 1
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T10:00:00Z"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T11:30:00Z"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T12:00:00Z"
					},
					"createdByUserId": {
						"type": "string",
						"example": "u1s2e3r4-i5d6-7890-1234-567890abcdef"
					},
					"updatedByUserId": {
						"type": "string",
						"example": "u1s2e3r4-i5d6-7890-1234-567890abcdef"
					},
					"companyId": {
						"type": "string",
						"example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
					},
					"company": {
						"$ref": "#/components/schemas/Company"
					},
					"userId": {
						"type": "string",
						"example": "u1s2e3r4-i5d6-7890-1234-567890abcdef"
					},
					"targetType": {
						"type": "string",
						"enum": [
							"Timesheet",
							"TimesheetEntry",
							"TimesheetApproval",
							"ActionCode"
						],
						"example": "Timesheet"
					},
					"targetId": {
						"type": "string",
						"example": "t1i2m3e4-s5h6e7e8-9012-3456-7890abcdef"
					},
					"action": {
						"type": "string",
						"enum": [
							"created",
							"updated",
							"submitted",
							"approved",
							"rejected",
							"deleted"
						],
						"example": "created"
					},
					"actorUserId": {
						"type": "string",
						"example": "a1c2t3o4-r5u6s7e8-9012-3456-7890abcdef"
					},
					"reason": {
						"type": "string",
						"example": "Initial creation"
					},
					"diff": {
						"$ref": "#/components/schemas/IStringToStringDictionary",
						"example": {
							"status": "DRAFT",
							"newStatus": "SUBMITTED"
						}
					},
					"metadata": {
						"$ref": "#/components/schemas/IStringToStringDictionary",
						"example": {
							"ipAddress": "192.168.1.1"
						}
					},
					"occurredAt": {
						"type": "string",
						"format": "date-time",
						"example": "2024-01-01T10:00:00Z"
					}
				},
				"required": [
					"id",
					"version",
					"createdAt",
					"updatedAt",
					"companyId",
					"company",
					"userId",
					"targetType",
					"targetId",
					"action",
					"occurredAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CompanySettings": {
				"properties": {
					"id": {
						"type": "string",
						"example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
					},
					"version": {
						"type": "number",
						"format": "double",
						"example": 1
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T10:00:00Z"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T11:30:00Z"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T12:00:00Z"
					},
					"createdByUserId": {
						"type": "string",
						"example": "u1s2e3r4-i5d6-7890-1234-567890abcdef"
					},
					"updatedByUserId": {
						"type": "string",
						"example": "u1s2e3r4-i5d6-7890-1234-567890abcdef"
					},
					"companyId": {
						"type": "string",
						"example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
					},
					"company": {
						"$ref": "#/components/schemas/Company"
					},
					"timezone": {
						"type": "string",
						"example": "UTC"
					},
					"workWeek": {
						"$ref": "#/components/schemas/Record_string.number-Array_",
						"example": {
							"monday": [
								9,
								17
							],
							"tuesday": [
								9,
								17
							]
						}
					},
					"holidayCalendar": {
						"type": "string",
						"example": "us_federal_holidays"
					},
					"timesheetApproverPolicy": {
						"$ref": "#/components/schemas/ApproverPolicy",
						"example": "manager_of_user"
					},
					"allowedEmailDomains": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"example": [
							"example.com",
							"another.org"
						]
					},
					"requireCompanyEmail": {
						"type": "boolean",
						"example": false
					},
					"defaultCountryCode": {
						"type": "string"
					},
					"defaultLocation": {
						"type": "string"
					},
					"officeCountryCodes": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"maxWeeklyMinutes": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"version",
					"createdAt",
					"updatedAt",
					"companyId",
					"company",
					"timezone",
					"workWeek",
					"timesheetApproverPolicy",
					"requireCompanyEmail",
					"defaultLocation",
					"maxWeeklyMinutes"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MenuCategory": {
				"properties": {
					"id": {
						"type": "string",
						"example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
					},
					"version": {
						"type": "number",
						"format": "double",
						"example": 1
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T10:00:00Z"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T11:30:00Z"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T12:00:00Z"
					},
					"createdByUserId": {
						"type": "string",
						"example": "u1s2e3r4-i5d6-7890-1234-567890abcdef"
					},
					"updatedByUserId": {
						"type": "string",
						"example": "u1s2e3r4-i5d6-7890-1234-567890abcdef"
					},
					"key": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"icon": {
						"type": "string"
					},
					"sortOrder": {
						"type": "number",
						"format": "double"
					},
					"companyId": {
						"type": "string"
					},
					"company": {
						"$ref": "#/components/schemas/Company"
					},
					"cards": {
						"items": {
							"$ref": "#/components/schemas/MenuCard"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"version",
					"createdAt",
					"updatedAt",
					"key",
					"title",
					"sortOrder",
					"companyId",
					"company",
					"cards"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LeaveType": {
				"enum": [
					"PTO",
					"SICK",
					"UNPAID"
				],
				"type": "string"
			},
			"LeaveRequestStatus": {
				"enum": [
					"PENDING",
					"APPROVED",
					"REJECTED",
					"CANCELLED"
				],
				"type": "string"
			},
			"LeaveRequest": {
				"properties": {
					"id": {
						"type": "string",
						"example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
					},
					"version": {
						"type": "number",
						"format": "double",
						"example": 1
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T10:00:00Z"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T11:30:00Z"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"example": "2023-10-27T12:00:00Z"
					},
					"createdByUserId": {
						"type": "string",
						"example": "u1s2e3r4-i5d6-7890-1234-567890abcdef"
					},
					"updatedByUserId": {
						"type": "string",
						"example": "u1s2e3r4-i5d6-7890-1234-567890abcdef"
					},
					"companyId": {
						"type": "string",
						"example": "a1b2c3d4-e5f6-7890-1234-567890abcdef"
					},
					"company": {
						"$ref": "#/components/schemas/Company"
					},
					"userId": {
						"type": "string",
						"example": "u1s2e3r4-i5d6-7890-1234-567890abcdef"
					},
					"user": {
						"$ref": "#/components/schemas/User"
					},
					"startDate": {
						"type": "string",
						"format": "date-time",
						"example": "2024-01-01"
					},
					"endDate": {
						"type": "string",
						"format": "date-time",
						"example": "2024-01-05"
					},
					"leaveType": {
						"$ref": "#/components/schemas/LeaveType",
						"example": "PTO"
					},
					"status": {
						"$ref": "#/components/schemas/LeaveRequestStatus",
						"example": "PENDING"
					},
					"reason": {
						"type": "string",
						"example": "Annual vacation trip"
					},
					"rejectionReason": {
						"type": "string",
						"example": "Insufficient coverage"
					}
				},
				"required": [
					"id",
					"version",
					"createdAt",
					"updatedAt",
					"companyId",
					"company",
					"userId",
					"user",
					"startDate",
					"endDate",
					"leaveType",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateLeaveRequestDto": {
				"properties": {
					"userId": {
						"type": "string",
						"example": "u1v2w3x4-y5z6-7890-1234-567890abcdef"
					},
					"startDate": {
						"type": "string",
						"example": "2024-01-01T00:00:00Z"
					},
					"endDate": {
						"type": "string",
						"example": "2024-01-05T00:00:00Z"
					},
					"leaveType": {
						"$ref": "#/components/schemas/LeaveType",
						"example": "vacation"
					},
					"reason": {
						"type": "string",
						"example": "Annual vacation trip"
					}
				},
				"required": [
					"userId",
					"startDate",
					"endDate",
					"leaveType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateLeaveRequestDto": {
				"properties": {
					"status": {
						"$ref": "#/components/schemas/LeaveRequestStatus",
						"example": "approved"
					},
					"rejectionReason": {
						"type": "string",
						"example": "Insufficient coverage"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"CreateCompanyDto": {
				"properties": {
					"name": {
						"type": "string",
						"example": "Acme Corp"
					},
					"timezone": {
						"type": "string",
						"example": "America/New_York"
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateCompanyDto": {
				"properties": {
					"name": {
						"type": "string",
						"example": "Acme Corporation"
					},
					"timezone": {
						"type": "string",
						"example": "America/Los_Angeles"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "gogotime-api",
		"version": "1.0.0",
		"description": "API for T-DEV-700 Project for Epitech",
		"contact": {
			"name": "Lazaro, Alexy, Massi, Lounis"
		}
	},
	"paths": {
...
		"/leave-requests": {
			"post": {
				"operationId": "CreateLeaveRequest",
				"responses": {
					"200": {
						"description": "The newly created leave request.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LeaveRequest"
								}
							}
						}
					}
				},
				"summary": "Creates a new leave request.",
				"tags": [
					"Leave Requests"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "The data for creating the leave request.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateLeaveRequestDto",
								"description": "The data for creating the leave request."
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllLeaveRequests",
				"responses": {
					"200": {
						"description": "An array of leave requests.",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/LeaveRequest"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"summary": "Retrieves all leave requests for the authenticated user's company.",
				"tags": [
					"Leave Requests"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/leave-requests/{id}": {
			"get": {
				"operationId": "GetLeaveRequest",
				"responses": {
					"200": {
						"description": "The leave request details.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LeaveRequest"
								}
							}
						}
					}
				},
				"summary": "Retrieves a single leave request by its ID.",
				"tags": [
					"Leave Requests"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The ID of the leave request to retrieve.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateLeaveRequest",
				"responses": {
					"200": {
						"description": "The updated leave request details.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LeaveRequest"
								}
							}
						}
					}
				},
				"summary": "Updates an existing leave request.",
				"tags": [
					"Leave Requests"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The ID of the leave request to update.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The data for updating the leave request.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateLeaveRequestDto",
								"description": "The data for updating the leave request."
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteLeaveRequest",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"summary": "Deletes a leave request by its ID.",
				"tags": [
					"Leave Requests"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The ID of the leave request to delete.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/companies": {
			"post": {
				"operationId": "CreateCompany",
				"responses": {
					"201": {
						"description": "Company created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Company"
								}
							}
						}
					}
				},
				"summary": "Creates a new company.",
				"tags": [
					"Companies"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "The data for creating the company.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateCompanyDto",
								"description": "The data for creating the company."
							}
						}
					}
				}
			}
		},
		"/companies/{id}": {
			"get": {
				"operationId": "GetCompany",
				"responses": {
					"200": {
						"description": "The company details.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Company"
								}
							}
						}
					}
				},
				"summary": "Retrieves a company by its ID.",
				"tags": [
					"Companies"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "The ID of the company to retrieve.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateCompany",
				"responses": {
					"200": {
						"description": "The updated company details.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Company"
								}
							}
						}
					}
				},
				"summary": "Updates an existing company.",
				"tags": [
					"Companies"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "The ID of the company to update.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The data for updating the company.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCompanyDto",
								"description": "The data for updating the company."
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}